fix: unhandled promise error in scheduleEmailReminders endpoint (#12183)
